Update autocompletion at appJS.js
Changes in lines 483 and 427

in line 483 i decided to change the aproach of using e.key to retrieve the last character that user typed to keep track of current word for autocompletion improve. i resolve  instead of that, creating a substring that goes from selectionstart -1 to selectionstart and then select the first character. that's 'cause in mobile i can't get e.key. then i change the event listener from keydown to keyup in line 427 'cause only after this event i can access to the last character typed. now i have pendding the task of track the exact x and y caret position to display correctlly the autocompletion list, 'cause it's still buggy